Refactored user registration vs tickets to be on its own concern
